test svd pairing
